+++ lib/modules/test/test_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: test
-// Updated at: 2026-07-15 16:01:17
-// Commit: chore(test): update dependencies and cleanup unused imports in test
+// Updated at: 2026-07-15 21:30:11
+// Commit: perf(test): apply RepaintBoundary on animated test widgets
 
 class testComponentState {
-  final String commitRef = '2857';
-  final String lastMessage = 'chore(test): update dependencies and cleanup unused imports in test';
-  final DateTime timestamp = DateTime.parse('2026-07-15T16:01:17+05:00');
+  final String commitRef = '2877';
+  final String lastMessage = 'perf(test): apply RepaintBoundary on animated test widgets';
+  final DateTime timestamp = DateTime.parse('2026-07-15T21:30:11+05:00');
 }